corrigiendo ejercicios

+++ Tema_3/Ejercicio2.java
@@ -11,7 +11,7 @@ public class Ejercicio2 {
       System.out.println("Diga los Euros que necesita saber lo que son en pesetas: ");
         double euro = Double.parseDouble(System.console().readLine()); //aquí creamos la variable para la introducción del primer número que el usuario a introducido por teclado
         
-        double total = (int) (euro * 166.386); //con esta variable hacemos el cálculo para pasar de Euros a Pesetas la cantidad deseada.
+        int total = (int) (euro * 166.386); //con esta variable hacemos el cálculo para pasar de Euros a Pesetas la cantidad deseada.
         
         System.out.println(euro + " Euros son:  " + total + " pesetas"); //mostramos el resultado de la conversion
